# Conflicts:
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/LCCData.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/directory/LCCKnowledgeData.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/asset/block/BombBoardBlockLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/AtomicBombBlockLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/DepositBlockLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/DungeonTableBlockLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/NetherReactorBlockLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/SaltBlockLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/SalvageBlockLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/TreetapBlockLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/TreetapContainerBlockLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/WastelandObeliskBlockLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/entity/ConsumerEntityLootFactory.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/generators/LCCLangData.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/generators/LCCLootData.kt
#	lcc-content-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/knowledge/ImageExport.kt
#	lcc-content/src/main/resources/assets/lcc/lang/en_gb.json
#	lcc-content/src/main/resources/assets/lcc/lang/en_us.json
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/DataAccessor.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/DataLauncher.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/batches/RecipeBatch.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/ChanceAlternativeBlockLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/ClusterBlockLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/DoorBlockLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/GenerousOreBlockLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/LeavesBlockLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/OreBlockLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/PottedPlantBlockLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/SelfBlockLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/SilkBlockLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/block/SimpleBlockLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/entity/ClassicEntityLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/entity/CustomEntityLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/loot/entity/FunctionalEntityLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/recipe/RecipeFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/factory/recipe/SlabLootFactory.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/json/recipe/RecipeStore.kt
#	lcc-data/src/main/kotlin/com/joshmanisdabomb/lcc/data/storage/RecipeStore.kt
#	lcc-data/src/main/resources/lcc-data.mixins.json

+++ lcc-content/src/main/kotlin/com/joshmanisdabomb/lcc/entity/render/ConsumerTongueEntityRenderer.kt
@@ -39,10 +39,10 @@ class ConsumerTongueEntityRenderer(context: EntityRendererFactory.Context) : Ent
         val d = MathHelper.lerp(tickDelta.toDouble(), owner.lastRenderX, owner.x) - MathHelper.lerp(tickDelta.toDouble(), origin.lastRenderX, origin.x)
         val e = entity.getTargetY(MathHelper.lerp(tickDelta.toDouble(), owner.lastRenderY, owner.y))!! - MathHelper.lerp(tickDelta.toDouble(), origin.lastRenderY, origin.y).plus(if (origin === hooked) hooked.height.div(2f) else 0f)
         val f = MathHelper.lerp(tickDelta.toDouble(), owner.lastRenderZ, owner.z) - MathHelper.lerp(tickDelta.toDouble(), origin.lastRenderZ, origin.z)
-        val g = MathHelper.sqrt(d * d + f * f).toDouble()
+        val g = MathHelper.sqrt((d * d + f * f).toFloat()).toDouble()
         val h = MathHelper.wrapDegrees((-(MathHelper.atan2(e, g) * 57.2957763671875)).toFloat())
         val i = MathHelper.wrapDegrees((MathHelper.atan2(f, d) * 57.2957763671875).toFloat() - 90.0f)
-        val j = MathHelper.sqrt(d * d + e * e + f * f)
+        val j = MathHelper.sqrt((d * d + e * e + f * f).toFloat())
         matrices.multiply(Vec3f.POSITIVE_Y.getDegreesQuaternion(180f - i))
         matrices.multiply(Vec3f.POSITIVE_X.getDegreesQuaternion(90f - h))
